Добавил пару элементов управления.

+++ ProductionLineApp/Statistics.h
@@ -4,9 +4,9 @@ struct Statistics
 {
 	unsigned long fstWorkplaceLoad;	// процент загрузки рабочих мест
 	unsigned long sndWorkplaceLoad;
-	unsigned long avgProdProcTime;	// в секундах (после запятой проценты миллисекунд, например 2,17 секунд)
-	int deferredProductsCount;
-	int fstWorkplaceAvgQueue;
-	int sndWorkplaceAvgQueue;
-	unsigned long fstWorkplaceBlockingTime;	// в секундах (после запятой проценты миллисекунд, например 2,17 секунд)
+	unsigned long avgProdProcTime;	// миллисекундах
+	unsigned long deferredProductsCount;
+	unsigned long fstWorkplaceAvgQueue;
+	unsigned long sndWorkplaceAvgQueue;
+	unsigned long fstWorkplaceBlockingTime;	// миллисекундах
 };
